modify

+++ models/catModel.js
@@ -6,7 +6,18 @@ const promisePool = pool.promise();
 const getAllCats = async (next) => {
   try {
     // TODO: do the LEFT (or INNER) JOIN to get owner's name as ownername (from wop_user table).
-    const [rows] = await promisePool.execute('SELECT * FROM wop_cat');
+    const [rows] = await promisePool.execute(`
+	SELECT 
+	cat_id, 
+	wop_cat.name, 
+	weight, 
+	owner, 
+	filename,
+	birthdate, 
+	wop_user.name as ownername 
+	FROM wop_cat 
+	JOIN wop_user ON 
+	wop_cat.owner = wop_user.user_id`);
     return rows;
   } catch (e) {
     console.error('getAllCats error', e.message);
@@ -17,7 +28,19 @@ const getAllCats = async (next) => {
 const getCat = async (id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'SELECT * FROM wop_cat WHERE cat_id = ?',
+      `
+	  SELECT 
+	  cat_id, 
+	  wop_cat.name, 
+	  weight, 
+	  owner, 
+	  filename,
+	  birthdate, 
+	  wop_user.name as ownername 
+	  FROM wop_cat 
+	  JOIN wop_user ON 
+	  wop_cat.owner = wop_user.user_id
+	  WHERE cat_id = ?`,
       [id]
     );
     return rows;
@@ -43,8 +66,8 @@ const addCat = async (name, weight, owner, birthdate, filename, next) => {
 const modifyCat = async (name, weight, owner, birthdate, cat_id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'UPDATE wop_cat SET name = ?, weight = ?, owner = ?, birthdate = ? WHERE cat_id = ?;',
-      [name, weight, owner, birthdate, cat_id]
+      'UPDATE wop_cat SET name = ?, weight = ?, birthdate = ? WHERE cat_id = ? AND owner = ?;',
+      [name, weight, birthdate, cat_id, owner]
     );
     return rows;
   } catch (e) {
@@ -53,11 +76,11 @@ const modifyCat = async (name, weight, owner, birthdate, cat_id, next) => {
   }
 };
 
-const deleteCat = async (id, next) => {
+const deleteCat = async (id, owner_id, next) => {
   try {
     const [rows] = await promisePool.execute(
-      'DELETE FROM wop_cat WHERE cat_id = ?',
-      [id]
+      'DELETE FROM wop_cat WHERE cat_id = ? AND owner = ?',
+      [id, owner_id]
     );
     return rows;
   } catch (e) {
